webapp/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
flake8==3.9.2
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-BasicAuth==0.2.0
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyMySQL==1.0.2
python-dotenv==0.18.0
requests==2.26.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
